+++ app/src/lib/build-info.js
@@ -1,2 +1,2 @@
-export const APP_BUILD = 'v1.4.1'
-export const BUILD_DATE = '2026-06-06 18:30'
+export const APP_BUILD = 'v1.4.2'
+export const BUILD_DATE = '2026-06-06 19:30'